create a BaseBillingUseCaseTest
