+++ purchases/src/test/java/com/revenuecat/purchases/google/usecase/QueryPurchaseHistoryUseCaseTest.kt
@@ -1,47 +1,26 @@
 package com.revenuecat.purchases.google.usecase
 
-import android.app.Activity
-import android.content.Intent
-import android.os.Handler
 import androidx.test.ext.junit.runners.AndroidJUnit4
-import com.android.billingclient.api.AcknowledgePurchaseParams
-import com.android.billingclient.api.AcknowledgePurchaseResponseListener
 import com.android.billingclient.api.BillingClient
-import com.android.billingclient.api.BillingClientStateListener
 import com.android.billingclient.api.BillingResult
 import com.android.billingclient.api.ConsumeParams
 import com.android.billingclient.api.ConsumeResponseListener
-import com.android.billingclient.api.ProductDetails
 import com.android.billingclient.api.PurchaseHistoryRecord
 import com.android.billingclient.api.PurchaseHistoryResponseListener
-import com.android.billingclient.api.PurchasesUpdatedListener
 import com.android.billingclient.api.QueryPurchaseHistoryParams
 import com.revenuecat.purchases.ProductType
 import com.revenuecat.purchases.PurchasesError
 import com.revenuecat.purchases.PurchasesErrorCode
-import com.revenuecat.purchases.common.BillingAbstract
-import com.revenuecat.purchases.common.DateProvider
-import com.revenuecat.purchases.common.caching.DeviceCache
-import com.revenuecat.purchases.common.diagnostics.DiagnosticsTracker
 import com.revenuecat.purchases.common.firstSku
-import com.revenuecat.purchases.google.BillingWrapper
 import com.revenuecat.purchases.google.toGoogleProductType
-import com.revenuecat.purchases.models.StoreProduct
-import com.revenuecat.purchases.utils.MockHandlerFactory
-import com.revenuecat.purchases.utils.mockProductDetails
 import com.revenuecat.purchases.utils.mockQueryPurchaseHistory
 import com.revenuecat.purchases.utils.verifyQueryPurchaseHistoryCalledWithType
-import io.mockk.Runs
-import io.mockk.clearAllMocks
 import io.mockk.every
-import io.mockk.just
 import io.mockk.mockk
 import io.mockk.slot
 import io.mockk.verify
 import org.assertj.core.api.Assertions.assertThat
 import org.assertj.core.api.Assertions.fail
-import org.junit.After
-import org.junit.Before
 import org.junit.Test
 import org.junit.runner.RunWith
 import org.robolectric.annotation.Config
@@ -52,112 +31,18 @@ import kotlin.time.Duration.Companion.milliseconds
 
 @RunWith(AndroidJUnit4::class)
 @Config(manifest = Config.NONE)
-class QueryPurchaseHistoryUseCaseTest {
+internal class QueryPurchaseHistoryUseCaseTest: BaseBillingUseCaseTest() {
 
     private companion object {
         const val timestamp0 = 1676379370000 // Tuesday, February 14, 2023 12:56:10.000 PM GMT
         const val timestamp123 = 1676379370123 // Tuesday, February 14, 2023 12:56:10.123 PM GMT
     }
 
-    private var onConnectedCalled: Boolean = false
-    private var mockClientFactory: BillingWrapper.ClientFactory = mockk()
-    private var mockClient: BillingClient = mockk()
-    private var purchasesUpdatedListener: PurchasesUpdatedListener? = null
-    private var billingClientStateListener: BillingClientStateListener? = null
-
-    private var mockDeviceCache: DeviceCache = mockk()
-    private var mockDiagnosticsTracker: DiagnosticsTracker = mockk()
-    private var mockDateProvider: DateProvider = mockk()
-
-    private var mockPurchasesListener: BillingAbstract.PurchasesUpdatedListener = mockk()
-
-    private var capturedAcknowledgeResponseListener = slot<AcknowledgePurchaseResponseListener>()
-    private var capturedAcknowledgePurchaseParams = slot<AcknowledgePurchaseParams>()
     private var capturedConsumeResponseListener = slot<ConsumeResponseListener>()
     private var capturedConsumeParams = slot<ConsumeParams>()
 
-    private lateinit var handler: Handler
-    private lateinit var wrapper: BillingWrapper
-
-    private lateinit var mockDetailsList: List<ProductDetails>
-
-    private var storeProducts: List<StoreProduct>? = null
-
-    private val billingClientOKResult = BillingClient.BillingResponseCode.OK.buildResult()
-    private var mockActivity = mockk<Activity>()
-
     private val subsGoogleProductType = ProductType.SUBS.toGoogleProductType()!!
     private val inAppGoogleProductType = ProductType.INAPP.toGoogleProductType()!!
-
-    @Before
-    fun setup() {
-        handler = MockHandlerFactory.createMockHandler()
-
-        storeProducts = null
-        purchasesUpdatedListener = null
-        billingClientStateListener = null
-
-        mockDiagnosticsTracker()
-        every { mockDateProvider.now } returns Date(1676379370000) // Tuesday, February 14, 2023 12:56:10 PM GMT
-
-        val listenerSlot = slot<PurchasesUpdatedListener>()
-        every {
-            mockClientFactory.buildClient(capture(listenerSlot))
-        } answers {
-            purchasesUpdatedListener = listenerSlot.captured
-            mockClient
-        }
-
-        val billingClientStateListenerSlot = slot<BillingClientStateListener>()
-        every {
-            mockClient.startConnection(capture(billingClientStateListenerSlot))
-        } answers {
-            billingClientStateListener = billingClientStateListenerSlot.captured
-        }
-
-        every {
-            mockClient.endConnection()
-        } just Runs
-
-        every {
-            mockClient.acknowledgePurchase(
-                capture(capturedAcknowledgePurchaseParams),
-                capture(capturedAcknowledgeResponseListener)
-            )
-        } just Runs
-
-        mockConsumeAsync(billingClientOKResult)
-
-        every {
-            mockClient.isReady
-        } returns false andThen true
-
-        val featureSlot = slot<String>()
-        every {
-            mockClient.isFeatureSupported(capture(featureSlot))
-        } returns billingClientOKResult
-
-        mockDetailsList = listOf(mockProductDetails())
-
-        wrapper = BillingWrapper(mockClientFactory, handler, mockDeviceCache, mockDiagnosticsTracker, mockDateProvider)
-        wrapper.purchasesUpdatedListener = mockPurchasesListener
-        wrapper.startConnectionOnMainThread()
-        onConnectedCalled = false
-        wrapper.stateListener = object : BillingAbstract.StateListener {
-            override fun onConnected() {
-                onConnectedCalled = true
-            }
-        }
-
-        every {
-            mockActivity.intent
-        } returns Intent()
-    }
-
-    @After
-    fun tearDown() {
-        clearAllMocks()
-    }
 
     @Test
     fun `if no listener is set, we fail`() {
@@ -180,8 +65,6 @@ class QueryPurchaseHistoryUseCaseTest {
 
     @Test
     fun `queryPurchaseHistoryAsync fails if sent invalid type`() {
-        billingClientStateListener!!.onBillingSetupFinished(billingClientOKResult)
-
         mockClient.mockQueryPurchaseHistory(
             billingClientOKResult,
             emptyList()
@@ -266,8 +149,6 @@ class QueryPurchaseHistoryUseCaseTest {
 
     @Test
     fun queryHistoryCallsListenerIfOk() {
-        billingClientStateListener!!.onBillingSetupFinished(billingClientOKResult)
-
         mockClient.mockQueryPurchaseHistory(
             billingClientOKResult,
             emptyList()
@@ -288,8 +169,6 @@ class QueryPurchaseHistoryUseCaseTest {
 
     @Test
     fun queryHistoryErrorCalledIfNotOK() {
-        billingClientStateListener!!.onBillingSetupFinished(billingClientOKResult)
-
         mockClient.mockQueryPurchaseHistory(
             BillingClient.BillingResponseCode.FEATURE_NOT_SUPPORTED.buildResult(),
             emptyList()
@@ -312,8 +191,6 @@ class QueryPurchaseHistoryUseCaseTest {
 
     @Test
     fun `queryPurchaseHistoryAsync sets correct type`() {
-        billingClientStateListener!!.onBillingSetupFinished(billingClientOKResult)
-
         val subsBuilder = mockClient.mockQueryPurchaseHistory(
             billingClientOKResult,
             emptyList()
@@ -650,9 +527,6 @@ class QueryPurchaseHistoryUseCaseTest {
 
     // endregion retries
 
-    private fun Int.buildResult(): BillingResult {
-        return BillingResult.newBuilder().setResponseCode(this).build()
-    }
 
     private fun mockConsumeAsync(billingResult: BillingResult) {
         every {
@@ -665,18 +539,4 @@ class QueryPurchaseHistoryUseCaseTest {
         }
     }
 
-    private fun mockDiagnosticsTracker() {
-        every {
-            mockDiagnosticsTracker.trackGoogleQueryProductDetailsRequest(any(), any(), any(), any())
-        } just Runs
-        every {
-            mockDiagnosticsTracker.trackGoogleQueryPurchasesRequest(any(), any(), any(), any())
-        } just Runs
-        every {
-            mockDiagnosticsTracker.trackGoogleQueryPurchaseHistoryRequest(any(), any(), any(), any())
-        } just Runs
-        every {
-            mockDiagnosticsTracker.trackProductDetailsNotSupported(any(), any())
-        } just Runs
-    }
 }
